put in a separate class
